requirements.txt
pandas==2.0.3          # For reading/writing Excel files
openpyxl==3.1.2        # Excel file support for pandas
google-generativeai==0.3.2  # Google Generative AI API for text and image generation
pillow==10.0.0         # Image processing (used with reportlab for PDFs)
reportlab==4.0.4       # PDF generation
pyyaml==6.0.1          # YAML configuration file parsing
python-dotenv==1.0.0   # Loading environment variables from .env file
streamlit==1.28.0
streamlit-option-menu==0.3.6
PyPDF2==3.0.1
